requirements.txt
setuptools~=57.0.0
pandas~=1.2.4
tqdm~=4.50.2
torch~=1.8.1+cpu
torchvision~=0.9.1+cpu
numpy~=1.20.3
scipy~=1.5.2
requests~=2.25.1
fire~=0.4.0
matplotlib~=3.5.1
pillow~=8.2.0
sty~=1.0.0rc1
cornet~=0.1.0
seaborn~=0.11.0
timm~=0.4.12
cloudpickle~=1.6.0
opencv-python~=4.5.1.48
filelock~=3.0.12